JRTB-4: added ability to send notifications about new articles JRTB-8: added ability to set inactive telegram user JRTB-9: added ability to set active user and/or start using it.

+++ src/main/java/com/github/javarushcommunity/javarushtelegrambot/repository/entity/GroupSub.java
@@ -16,13 +16,13 @@ import static java.util.Objects.isNull;
 public class GroupSub {
 
     @Id
-    private Integer id;
+    private Long id;
 
     @Column(name = "title")
     private String title;
 
     @Column(name = "last_article_id")
-    private Integer lastArticleId;
+    private Long lastArticleId;
 
     @ManyToMany(fetch = FetchType.EAGER)
     @JoinTable(
